feat: working in-dashboard settings
