+++ package.json
@@ -24,7 +24,7 @@
 		"clsx": "^2.1.1",
 		"date-fns": "^4.1.0",
 		"dompurify": "^3.2.6",
-		"framer-motion": "^12.9.2",
+		"framer-motion": "^12.23",
 		"lucide-react": "^0.456.0",
 		"marked": "^15.0.7",
 		"monaco-editor": "^0.52.2",
