--- src/components/settings/old/published-scripts.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-import type React from 'react';
-import { Primary } from '../../ui/Buttons';
-
-export const PublishedScripts: React.FC = () => {
-	return (
-		<Primary
-			Title="Add Created/Published Script"
-			onClick={() => console.log('Add Created/Published Script')}
-		/>
-	);
-};
